pyproject.toml
[project]
name = "feedback-system"
version = "0.2.0"
description = "AI-Powered Customer Feedback Analysis System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.20.0",
    "asyncpg>=0.29.0",
    "apscheduler>=3.10",
    "beautifulsoup4>=4.12",
    "fastapi>=0.129.2",
    "groq>=0.9.0",
    "httpx>=0.27",
    "matplotlib>=3.9",
    "numpy<2",
    "pandas<3",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.22",
    "reportlab>=4.2",
    "streamlit>=1.35",
    "plotly>=5.22",
    "torch==2.0.1",
    "tortoise-orm>=1.1.5",
    "transformers==4.30.0",
    "uvicorn>=0.41.0",
    "google-play-scraper>=1.2.7",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.15.2",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "dashboard*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
